1440: Allow Consent API controllers to be configured for specific OB versions
Refactoring the DomesticPaymentConsentApi, removing the class level @RequestMapping annotation from the interface and instead supplying it in the version specific controller impl.

Extracting BaseDomesticPaymentConsentApiController from DomesticPaymentConsentApiController, this class contains the common controller logic that is the same across versions. Version specific controllers extend the Base and supply the OBVersion parameter as well as the annotations to wire the controller in Spring and to specify the @RequestMapping (as the URL pattern is version specific).

The same pattern will be followed for all Consent API controllers.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1440
